+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # <img src="https://uploads-ssl.webflow.com/5ea5d3315186cf5ec60c3ee4/5edf1c94ce4c859f2b188094_logo.svg" alt="Pip.Services Logo" width="200"> <br/> Component Definitions for .NET Changelog
 
+
+## <a name="3.2.3"></a> 3.2.3 (2021-07-23) 
+
+### Features
+* **ConnectionParams** added FromTuples
+* **test** Added DefaultTestFactory class
+* **test** Added Shutdown class
+
+
 ## <a name="3.2.2"></a> 3.2.2 (2021-05-24) 
 
 ### Bug Fixes